refactor(ai): extract observer_goal_phase cases and residual fat-pin gate
Streamline the Phase-8 residual observer_goal_phase pin suite into topical
cases modules and enforce the ≤750-or-cases ceiling via repo lint.
